@@ -56,52 +56,52 @@
     "html-webpack-plugin": "^3.0.7",
     "jest": "^22.4.3",
     "json-loader": "^0.5.4",
-    "mini-css-extract-plugin": "^0.4.0",
+    "mini-css-extract-plugin": "^0.4.5",
     "mock-socket": "^7.1.0",
     "moment": "^2.22.1",
     "numeral": "^2.0.6",
     "postcss-loader": "^1.2.2",
     "prop-types": "^15.6.1",
-    "qrcode": "^1.2.0",
-    "query-string": "^6.0.0",
+    "qrcode": "^1.3.3",
+    "query-string": "^6.2.0",
     "raw-loader": "^0.5.1",
-    "react": "^16.2.0",
-    "react-chartjs-2": "^2.7.2",
-    "react-datetime": "^2.14.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.8.2",
+    "react-chartjs-2": "^2.7.4",
+    "react-datetime": "^2.16.3",
+    "react-dom": "^16.8.2",
     "react-helmet": "^5.2.0",
-    "react-hot-loader": "^4.0.0",
+    "react-hot-loader": "^4.6.5",
     "react-json-tree": "^0.11.2",
     "react-json-view": "^1.19.1",
     "react-loading": "^1.0.4",
-    "react-modal": "^3.3.2",
-    "react-popper": "^1.3.2",
-    "react-redux": "^5.0.7",
-    "react-redux-loading-bar": "^4.0.5",
+    "react-modal": "^3.8.1",
+    "react-popper": "^1.3.3",
+    "react-redux": "^5.1.1",
+    "react-redux-loading-bar": "^4.2.0",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
-    "react-transition-group": "^2.3.0",
+    "react-transition-group": "^2.5.3",
     "recompose": "^0.26.0",
     "redux": "^3.7.2",
-    "redux-devtools-extension": "^2.13.2",
+    "redux-devtools-extension": "^2.13.8",
     "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.2.0",
     "reselect": "^3.0.1",
     "reset-css": "^3.0.0",
     "style-loader": "^0.20.3",
-    "styled-components": "^3.2.3",
-    "styled-reset": "^1.3.2",
-    "stylelint": "^9.1.3",
+    "styled-components": "^3.4.10",
+    "styled-reset": "^1.6.8",
+    "stylelint": "^9.10.1",
     "stylelint-config-recommended": "^2.1.0",
     "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-processor-styled-components": "^1.3.1",
+    "stylelint-processor-styled-components": "^1.5.2",
     "url-join": "^4.0.0",
-    "url-loader": "^1.0.1",
+    "url-loader": "^1.1.2",
     "uuid": "^3.2.1",
-    "webpack": "^4.5.0",
+    "webpack": "^4.29.4",
     "webpack-bundle-analyzer": "^2.11.1",
-    "webpack-cli": "^2.0.14",
-    "webpack-dev-server": "^3.1.3"
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.1.14"
   },
   "devDependencies": {
     "babel-core": "^7.0.0-0",
