@@ -1,15 +1,7 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
-import SideNavigation from './SideNavigation'
-import TopBar from './TopBar'
+import SideNavigation from './SideNavigationV2'
 import PropTypes from 'prop-types'
-import AccountSelectorMenu from './AccountSelectorMenu'
-import withClickOutsideEnhancer from '../enhancer/clickOutside'
-import { compose } from 'recompose'
-import AccountsFetcher from '../omg-account/accountsFetcher'
-import { connect } from 'react-redux'
-import { withRouter } from 'react-router-dom'
-import { switchAccount } from '../omg-account-current/action'
 import Alert from '../omg-alert'
 import LoadingBar from 'react-redux-loading-bar'
 import TransactionRequestPanel from '../omg-transaction-request-tab'
@@ -26,7 +18,7 @@ const SideNav = styled(SideNavigation)`
   display: inline-block;
   vertical-align: top;
   flex: 0 0 auto;
-  width: 240px;
+  width: 220px;
 `
 
 const ContentContainer = styled.div`
@@ -38,77 +30,12 @@ const ContentContainer = styled.div`
 const Content = styled.div`
   padding: 0 7%;
 `
-const enhance = compose(
-  connect(
-    null,
-    { switchAccount }
-  ),
-  withRouter,
-  withClickOutsideEnhancer
-)
-const EnhancedAccountSelectorMenuClickOutside = enhance(
-  class AccountSelectorMenuClickOutside extends Component {
-    static propTypes = {
-      closeSwitchAccountTab: PropTypes.func,
-      location: PropTypes.object,
-      history: PropTypes.object,
-      switchAccount: PropTypes.func
-    }
-    state = {
-      searchValue: ''
-    }
-    handleClickOutside = () => {
-      this.props.closeSwitchAccountTab()
-    }
-    onKeyDown = e => {
-      if (e.keyCode === 27) this.props.closeSwitchAccountTab()
-    }
-    onClickAccountItem = account => e => {
-      this.props.history.push(`/${account.id}/dashboard`)
-      this.handleClickOutside()
-      this.props.switchAccount(account)
-    }
-
-    onSearchChange = e => {
-      this.setState({ searchValue: e.target.value })
-    }
-
-    render () {
-      return (
-        <AccountsFetcher
-          query={{ search: this.state.searchValue, perPage: 20, page: 1 }}
-          render={({ data: accounts }) => {
-            return (
-              <AccountSelectorMenu
-                accounts={accounts}
-                onClickAccountItem={this.onClickAccountItem}
-                onKeyDown={this.onKeyDown}
-                onSearchChange={this.onSearchChange}
-                searchValue={this.state.searchValue}
-              />
-            )
-          }}
-        />
-      )
-    }
-  }
-)
-
 class AppLayout extends Component {
   static propTypes = {
     children: PropTypes.node,
     location: PropTypes.object
   }
-  state = {
-    switchAccount: false
-  }
 
-  closeSwitchAccountTab = () => {
-    this.setState({ switchAccount: false })
-  }
-  onClickSwitchAccount = () => {
-    this.setState({ switchAccount: true })
-  }
   scrollTopContentContainer = () => {
     this.contentContainer.scrollTo(0, 0)
   }
@@ -117,17 +44,8 @@ class AppLayout extends Component {
     return (
       <Container>
         <LoadingBar updateTime={1000} style={{ backgroundColor: '#1A56F0', zIndex: 99999 }} />
-        <SideNav
-          switchAccount={this.state.switchAccount}
-          onClickSwitchAccount={this.onClickSwitchAccount}
-        />
-        {this.state.switchAccount && (
-          <EnhancedAccountSelectorMenuClickOutside
-            closeSwitchAccountTab={this.closeSwitchAccountTab}
-          />
-        )}
+        <SideNav />
         <ContentContainer innerRef={contentContainer => (this.contentContainer = contentContainer)}>
-          <TopBar />
           <Content>
             {React.cloneElement(this.props.children, {
               scrollTopContentContainer: this.scrollTopContentContainer
