@@ -87,6 +87,11 @@ export const cacheReducer = createReducer(
         [action.cacheKey]: { ids: action.data.map(d => d.id), pagination: action.pagination }
       }
     },
-    'CURRENT_ACCOUNT/SWITCH': () => ({})
+    'ADMINS/REQUEST/SUCCESS': (state, action) => {
+      return {
+        ...state,
+        [action.cacheKey]: { ids: action.data.map(d => d.id), pagination: action.pagination }
+      }
+    }
   }
 )
