@@ -7,7 +7,7 @@ import { createTransactionRequest } from '../omg-transaction-request/action'
 import { connect } from 'react-redux'
 import TokensFetcher from '../omg-token/tokensFetcher'
 import WalletsFetcher from '../omg-wallet/walletsFetcher'
-import { selectPrimaryWalletCurrentAccount } from '../omg-wallet/selector'
+import { selectPrimaryWalletByAccountId } from '../omg-wallet/selector'
 import { withRouter } from 'react-router-dom'
 import { compose } from 'recompose'
 import { formatAmount } from '../utils/formatter'
@@ -125,7 +125,9 @@ const CollapsableContent = styled.div`
 const enhance = compose(
   withRouter,
   connect(
-    state => ({ primaryWallet: selectPrimaryWalletCurrentAccount(state) }),
+    state => ({
+      primaryWallet: selectPrimaryWalletByAccountId(state)(this.props.match.params.accountId)
+    }),
     { createTransactionRequest }
   )
 )
@@ -160,7 +162,9 @@ class CreateTransactionRequest extends Component {
         amount: formatAmount(this.state.amount, _.get(this.state.selectedToken, 'subunit_to_unit')),
         tokenId: _.get(this.state, 'selectedToken.id'),
         address: this.state.address || _.get(this.props, 'primaryWallet.address'),
-        accountId: _.get(this.state, 'selectedWallet.account_id', this.props.match.params.accountId),
+        accountId:
+          _.get(this.state, 'selectedWallet.account_id') ||
+          _.get(this.props, 'march.params.accountId'),
         expirationDate: this.state.expirationDate && moment(this.state.expirationDate).toISOString()
       })
       if (result.data) {
