@@ -73,7 +73,7 @@ class LoginForm extends Component {
         })
         if (result.data) {
           this.setState({ error: null })
-          this.props.history.push(`/${result.data.account_id}/dashboard`)
+          this.props.history.push('/')
         } else {
           this.setState({ error: result.error.description || result.error.message, submitted: false })
         }
