@@ -23,13 +23,26 @@ const ConsumptionPageContainer = styled.div`
   td {
     white-space: nowrap;
   }
-  td:nth-child(3) {
+  td:first-child{
+    border: none;
+    position: relative;
+    :before {
+      content: '';
+      position: absolute;
+      right: 0;
+      bottom: -1px;
+      height: 1px;
+      width: calc(100% - 50px);
+      border-bottom: 1px solid ${props => props.theme.colors.S100};
+    }
+  }
+  td:nth-child(2) {
     width: 150px;
     max-width: 150px;
     overflow: hidden;
     text-overflow: ellipsis;
   }
-  td:nth-child(6) {
+  td:nth-child(5) {
     text-transform: capitalize;
   }
   tr:hover {
@@ -39,7 +52,7 @@ const ConsumptionPageContainer = styled.div`
       }
     }
   }
-  i[name="Copy"] {
+  i[name='Copy'] {
     margin-left: 5px;
     cursor: pointer;
     visibility: hidden;
@@ -51,8 +64,11 @@ const ConsumptionPageContainer = styled.div`
 `
 const SortableTableContainer = styled.div`
   position: relative;
-  i[name="Consumption"] {
-    color: ${props => props.theme.colors.BL400};
+  i[name='Consumption'] {
+    color: ${props => props.theme.colors.B100};
+    padding: 8px;
+    border-radius: 6px;
+    border: 1px solid ${props => props.theme.colors.S400};
   }
 `
 export const NameColumn = styled.div`
@@ -65,17 +81,18 @@ class ConsumptionPage extends Component {
     match: PropTypes.object,
     history: PropTypes.object,
     location: PropTypes.object,
-    scrollTopContentContainer: PropTypes.func
+    scrollTopContentContainer: PropTypes.func,
+    query: PropTypes.object,
+    fetcher: PropTypes.node
+  }
+  static defaultProps = {
+    query: {},
+    fetcher: ConsumptionFetcher
   }
   constructor (props) {
     super(props)
-    this.state = {
-      createAccountModalOpen: false,
-      exportModalOpen: false,
-      loadMoreTime: 1
-    }
     this.columns = [
-      { key: 'id', title: 'REQUEST ID', sort: true },
+      { key: 'id', title: 'CONSUMPTION ID', sort: true },
       { key: 'type', title: 'TYPE', sort: true },
       { key: 'estimated_consumption_amount', title: 'AMOUNT', sort: true },
       { key: 'created_by', title: 'CONSUMER' },
@@ -132,7 +149,10 @@ class ConsumptionPage extends Component {
     return (
       <ConsumptionPageContainer>
         <TopNavigation title={'Transaction Consumptions'} buttons={[]} />
-        <SortableTableContainer innerRef={table => (this.table = table)} loadingStatus={individualLoadingStatus}>
+        <SortableTableContainer
+          innerRef={table => (this.table = table)}
+          loadingStatus={individualLoadingStatus}
+        >
           <SortableTable
             rows={consumptions}
             columns={this.columns}
@@ -145,26 +165,22 @@ class ConsumptionPage extends Component {
             activeIndexKey={activeIndexKey}
           />
         </SortableTableContainer>
-        <CreateAccountModal
-          open={this.state.createAccountModalOpen}
-          onRequestClose={this.onRequestCloseCreateAccount}
-          onCreateAccount={fetch}
-        />
-        <ExportModal open={this.state.exportModalOpen} onRequestClose={this.onRequestCloseExport} />
       </ConsumptionPageContainer>
     )
   }
 
   render () {
+    const Fetcher = this.props.fetcher
     return (
-      <ConsumptionFetcher
+      <Fetcher
         render={this.renderConsumptionPage}
         {...this.state}
         {...this.props}
         query={{
           page: queryString.parse(this.props.location.search).page,
           perPage: Math.floor(window.innerHeight / 65),
-          searchTerms: {id: queryString.parse(this.props.location.search).search}
+          searchTerms: { id: queryString.parse(this.props.location.search).search },
+          ...this.props.query
         }}
         onFetchComplete={this.props.scrollTopContentContainer}
       />
