@@ -23,13 +23,15 @@ const TokenDetailPageContainer = styled.div`
   td:first-child {
     width: 50%;
   }
-  td:nth-child(2),
-  td:nth-child(3) {
+  td:nth-child(3),
+  td:nth-child(4) {
     width: 25%;
   }
 `
 const columns = [
+
   { key: 'token', title: 'TOKEN NAME', sort: true },
+  { key: 'id', title: 'TOKEN ID', sort: true },
   { key: 'symbol', title: 'SYMBOL', sort: true },
   { key: 'created', title: 'CREATED DATE', sort: true }
 ]
@@ -117,8 +119,7 @@ class TokenDetailPage extends Component {
     return data
   }
   onClickRow = (data, index) => e => {
-    const { params } = this.props.match
-    this.props.history.push(`/${params.accountId}/tokens/${data.id}`)
+    this.props.history.push(`/tokens/${data.id}`)
   }
   renderTokenDetailPage = ({ data: tokens, individualLoadingStatus, pagination, fetch }) => {
     const data = tokens.map(token => {
