@@ -6,7 +6,6 @@ import UserProvider from '../omg-users/userProvider'
 import { compose } from 'recompose'
 import Section, { DetailGroup } from '../omg-page-detail-layout/DetailSection'
 import TopBar from '../omg-page-detail-layout/TopBarDetail'
-import DetailLayout from '../omg-page-detail-layout/DetailLayout'
 import moment from 'moment'
 import { LoadingSkeleton } from '../omg-uikit'
 import { formatReceiveAmountToTotal } from '../utils/formatter'
@@ -71,16 +70,13 @@ class TokenDetailPage extends Component {
     )
   }
   renderWallet = wallet => {
-    console.log(wallet)
-    const accountId = this.props.match.params.accountId
     return (
       <Section title='BALANCE'>
         {wallet ? (
           <div>
             <DetailGroup>
-              <b>Wallet Address:</b>{' '}
-              <Link to={`/${accountId}/wallets/${wallet.address}`}>{wallet.address}</Link> ({' '}
-              <span>{wallet.name}</span> )
+              <b>Wallet Address:</b> <Link to={`/wallets/${wallet.address}`}>{wallet.address}</Link>{' '}
+              ( <span>{wallet.name}</span> )
             </DetailGroup>
             {wallet.balances.map(balance => {
               return (
@@ -105,17 +101,14 @@ class TokenDetailPage extends Component {
     )
   }
   renderUserDetailContainer = (user, wallet) => {
-    const accountId = this.props.match.params.accountId
     return (
-      <DetailLayout backPath={`/${accountId}/users`}>
-        <ContentContainer>
-          {this.renderTopBar(user)}
-          <ContentDetailContainer>
-            <DetailContainer>{this.renderDetail(user)}</DetailContainer>
-            {wallet && <DetailContainer>{this.renderWallet(wallet)}</DetailContainer>}
-          </ContentDetailContainer>
-        </ContentContainer>
-      </DetailLayout>
+      <ContentContainer>
+        {this.renderTopBar(user)}
+        <ContentDetailContainer>
+          <DetailContainer>{this.renderDetail(user)}</DetailContainer>
+          {wallet && <DetailContainer>{this.renderWallet(wallet)}</DetailContainer>}
+        </ContentDetailContainer>
+      </ContentContainer>
     )
   }
 
