@@ -21,6 +21,8 @@ const UserPageContainer = styled.div`
   }
   td:first-child {
     width: 40%;
+    border: none;
+    position: relative;
   }
   tr:hover {
     td:nth-child(1) {
@@ -50,9 +52,12 @@ const UserIdContainer = styled.div`
   span {
     vertical-align: middle;
   }
-  i {
+  i[name='Profile'] {
     margin-right: 5px;
-    color: ${props => props.theme.colors.BL400};
+    color: ${props => props.theme.colors.B100};
+    padding: 8px;
+    border-radius: 6px;
+    border: 1px solid ${props => props.theme.colors.S400};
   }
 `
 class UsersPage extends Component {
@@ -60,7 +65,15 @@ class UsersPage extends Component {
     location: PropTypes.object,
     history: PropTypes.object,
     match: PropTypes.object,
-    scrollTopContentContainer: PropTypes.func
+    scrollTopContentContainer: PropTypes.func,
+    query: PropTypes.object,
+    fetcher: PropTypes.func,
+    onClickRow: PropTypes.func
+  }
+
+  static defaultProps = {
+    query: {},
+    fetcher: UsersFetcher
   }
   constructor (props) {
     super(props)
@@ -70,8 +83,7 @@ class UsersPage extends Component {
     }
   }
   onClickRow = (data, index) => e => {
-    const { params } = this.props.match
-    this.props.history.push(`/${params.accountId}/users/${data.id}`)
+    this.props.history.push(`/users/${data.id}`)
   }
   renderExportButton = () => {
     return (
@@ -90,7 +102,7 @@ class UsersPage extends Component {
   }
   getColumns = () => {
     return [
-      { key: 'id', title: 'ID', sort: true },
+      { key: 'id', title: 'USER ID', sort: true },
       { key: 'email', title: 'EMAIL', sort: true },
       { key: 'username', title: 'USERNAME', sort: true },
       { key: 'created_at', title: 'CREATED DATE', sort: true },
@@ -134,7 +146,7 @@ class UsersPage extends Component {
             columns={this.getColumns(users)}
             loadingStatus={individualLoadingStatus}
             rowRenderer={this.rowRenderer}
-            onClickRow={this.onClickRow}
+            onClickRow={this.props.onClickRow || this.onClickRow}
             isFirstPage={pagination.is_first_page}
             isLastPage={pagination.is_last_page}
             navigation
@@ -147,16 +159,17 @@ class UsersPage extends Component {
   }
 
   render () {
+    const Fetcher = this.props.fetcher
     return (
-      <UsersFetcher
+      <Fetcher
         {...this.state}
         {...this.props}
         render={this.renderUserPage}
         query={{
           page: queryString.parse(this.props.location.search).page,
           perPage: 15,
-          accountId: this.props.match.params.accountId,
-          ...createSearchUsersQuery(queryString.parse(this.props.location.search).search)
+          ...createSearchUsersQuery(queryString.parse(this.props.location.search).search),
+          ...this.props.query
         }}
         onFetchComplete={this.props.scrollTopContentContainer}
       />
