@@ -10,12 +10,20 @@ const AvatarCircle = styled.div`
   background-size: cover;
   vertical-align: middle;
   background-position: center;
-  color: ${props => props.theme.colors.BL400};
   border-radius: 4px;
   font-weight: 600;
   text-align: center;
-  line-height: 30px;
   font-size: 8px;
+  position: relative;
+  > div {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    margin: 0 auto;
+    left: 0;
+    right: 0;
+    color: ${props => props.theme.colors.BL400};
+  }
 `
 export default class Avatar extends Component {
   static propTypes = {
@@ -27,10 +35,14 @@ export default class Avatar extends Component {
     name: ''
   }
   render () {
-    const name = _.isNil(this.props.name) ? '' : String(this.props.name).substring(0, 3).toUpperCase()
+    const name = _.isNil(this.props.name)
+      ? ''
+      : String(this.props.name)
+          .substring(0, 3)
+          .toUpperCase()
     return (
       <AvatarCircle image={this.props.image} size={this.props.size} {...this.props}>
-        {this.props.image ? '' : name}
+        <div>{this.props.image ? '' : name}</div>
       </AvatarCircle>
     )
   }
