@@ -1,9 +1,14 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
+import Icon from '../icon'
 const BreadcrumbContainer = styled.div`
   position: relative;
   color: ${props => props.theme.colors.B100};
+  i {
+    padding: 0 5px;
+    font-size: inherit;
+  }
 `
 export default class Breadcrumb extends Component {
   static propTypes = {
@@ -11,11 +16,12 @@ export default class Breadcrumb extends Component {
   }
 
   render () {
+    const items = this.props.items.filter(x => x)
     return (
       <BreadcrumbContainer>
-        {this.props.items.reduce((prev, curr, index) => {
+        {items.reduce((prev, curr, index) => {
           prev.push(curr)
-          if (index < this.props.items.length - 1) prev.push(' > ')
+          if (index < items.length - 1) prev.push(<Icon name='Chevron-Right' />)
           return prev
         }, [])}
       </BreadcrumbContainer>
