@@ -86,7 +86,7 @@ class Table extends Component {
   renderLoadingRows = () => {
     return this.loadingWidthBars.map((x, i) => {
       return (
-        <tr key={`row-${i}`} ref={row => (this.row = row)}>
+        <tr key={`row-${i}`} ref={row => (this.row = row)} style={{ pointerEvents: 'none' }}>
           <td key={`col-rest-${i}`}>
             <LoadingSkeleton height={'12px'} width={x} style={{ margin: '5px 0' }} />
           </td>
@@ -128,13 +128,13 @@ class Table extends Component {
         height={this.props.loadingRowNumber * 40}
         loading={this.props.loading}
       >
-        <Fade in={this.props.loading} timeout={300} key={'loading'} unmountOnExit>
+        <Fade in={this.props.loading} timeout={200} key={'loading'} unmountOnExit>
           <table style={{ position: 'absolute', background: 'white' }}>
             <thead>{this.renderLoadingColumns()}</thead>
             <tbody>{this.renderLoadingRows()}</tbody>
           </table>
         </Fade>
-        <Fade in={!this.props.loading} timeout={300} key={'data'} unmountOnExit appear>
+        <Fade in={!this.props.loading} timeout={200} key={'data'} unmountOnExit appear>
           <table>
             <thead>{this.renderColumns()}</thead>
             {!!dataRows.length && <tbody>{this.renderDataRows()}</tbody>}
