@@ -1,5 +1,5 @@
 import { createSelectAllPagesCachedQuery } from '../omg-cache/selector'
-import { selectCurrentAccount } from '../omg-account-current/selector'
+import { selectGetAccountById } from '../omg-account/selector'
 import _ from 'lodash'
 export const selectWallets = (state, search) => {
   return _.values(state.wallets).filter(wallet => {
@@ -13,9 +13,9 @@ export const selectWalletsByAccountId = (state, search, accountId) => {
     return wallet.account_id === accountId && (reg.test(wallet.name) || reg.test(wallet.address))
   })
 }
-export const selectPrimaryWalletCurrentAccount = state => {
+export const selectPrimaryWalletByAccountId = state => accountId => {
   return _.values(state.wallets).find(wallet => {
-    const account = selectCurrentAccount(state)
+    const account = selectGetAccountById(state)(accountId)
     return wallet.account_id === account.id && wallet.identifier === 'primary'
   })
 }
