@@ -1,9 +1,9 @@
-import { selectPrimaryWalletCurrentAccount } from './selector'
+import { selectPrimaryWalletByAccountId } from './selector'
 
 describe('selectors wallet', () => {
   test('should select primary wallet of current account', () => {
     const state = {
-      currentAccount: { id: 'a' },
+      accounts: { id: 'a' },
       wallets: {
         b: {
           account_id: 'a',
@@ -15,7 +15,7 @@ describe('selectors wallet', () => {
         }
       }
     }
-    expect(selectPrimaryWalletCurrentAccount(state)).toEqual({
+    expect(selectPrimaryWalletByAccountId(state)('a')).toEqual({
       account_id: 'a',
       identifier: 'primary'
     })
