@@ -1,5 +1,5 @@
 import { authenticatedRequest } from './apiService'
-export function getConsumptions ({ perPage, page, sort, search, searchTerms }) {
+export function getConsumptions ({ perPage, page, sort, search, searchTerms, matchAll, matchAny }) {
   return authenticatedRequest({
     path: '/transaction_consumption.all',
     data: {
@@ -8,7 +8,9 @@ export function getConsumptions ({ perPage, page, sort, search, searchTerms }) {
       sort_by: sort.by,
       sort_dir: sort.dir,
       search_term: search,
-      search_terms: searchTerms
+      search_terms: searchTerms,
+      match_all: matchAll,
+      match_any: matchAny
     }
   })
 }
