@@ -1,6 +1,6 @@
 import { authenticatedRequest } from './apiService'
 import uuid from 'uuid/v4'
-export function getTransactionRequests ({ perPage, page, sort, search }) {
+export function getTransactionRequests ({ perPage, page, sort, search, matchAll, matchAny }) {
   return authenticatedRequest({
     path: '/transaction_request.all',
     data: {
@@ -8,7 +8,9 @@ export function getTransactionRequests ({ perPage, page, sort, search }) {
       page: Number(page) || 1,
       sort_by: sort.by,
       sort_dir: sort.dir,
-      search_term: search
+      search_term: search,
+      match_all: matchAll,
+      match_any: matchAny
     }
   })
 }
